reg save

+++ stm32/jt/STM32CubeIDE/Application/User/gui/JtPresenter.cpp
@@ -63,7 +63,7 @@ uint16_t JtPresenter::getWeldingTimeSec()
 
 float JtPresenter::getWeldingAmps()
 {
-	uint32_t res = 0;
+	float res = 0.0;
 	res = model->getWeldingAmps();
 	return res;
 }
